fix(uploads): use a Uint32Array skip table in the multipart parser
Supersedes the boundary-length cap from the streaming-parser commit. The
root cause is the Uint8Array skip table in buffer-search.ts: for a boundary
of 252+ characters the search needle is 256 bytes and the fill wraps to 0,
so the search loops forever. Our parser is Bun-only and always uses this JS
path. Uint32Array can't wrap for any needle a JavaScript string can produce,
fixes it with no behavior change, and lets the constructor drop the RFC 2046
length guard.

Matches the fix proposed upstream (GHSA-hpwp-7w8r-373j).

+++ apps/api/src/lib/multipart/buffer-search.ts
@@ -6,9 +6,10 @@ export type SearchFunction = (haystack: Uint8Array, start?: number) => number;
 export function createSearch(pattern: string): SearchFunction {
     const needle = new TextEncoder().encode(pattern);
     const needleEnd = needle.length - 1;
-    // Skip distances must fit in a byte; callers keep the needle short by capping the
-    // boundary length (see MAX_BOUNDARY_LENGTH), so a distance can never wrap to 0.
-    const skipTable = new Uint8Array(256).fill(needle.length);
+    // Uint32, not Uint8: for a needle of 256+ bytes a Uint8 fill would wrap to 0, making
+    // the default skip 0 so the search never advances (infinite loop). needle.length is
+    // bounded by the pattern string length, well under 2^32, so a Uint32 skip can't wrap.
+    const skipTable = new Uint32Array(256).fill(needle.length);
     for (let i = 0; i < needleEnd; ++i) {
         skipTable[needle[i]] = needleEnd - i;
     }
